+++ packages/web/app/src/pages/organization-new.tsx
@@ -1,27 +1,44 @@
 import { ReactElement } from 'react';
 import { LogOutIcon } from 'lucide-react';
+import { useForm, UseFormReturn } from 'react-hook-form';
+import { useMutation } from 'urql';
+import { z } from 'zod';
 import { Button } from '@/components/ui/button';
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardFooter,
+  CardHeader,
+  CardTitle,
+} from '@/components/ui/card';
 import { DottedBackground } from '@/components/ui/dotted-background';
+import { Form, FormControl, FormField, FormItem, FormMessage } from '@/components/ui/form';
+import { Input } from '@/components/ui/input';
 import { Meta } from '@/components/ui/meta';
-import { CreateOrganizationForm } from '@/components/ui/modals/create-organization';
+import { Spinner } from '@/components/ui/spinner';
+import { useToast } from '@/components/ui/use-toast';
 import { HiveLogo } from '@/components/v2/icon';
+import { graphql } from '@/gql';
+import { zodResolver } from '@hookform/resolvers/zod';
 import { Link, useRouter } from '@tanstack/react-router';
 
 export function NewOrgPage(): ReactElement {
   const router = useRouter();
-
-  const handleOnClick = (): void => {
-    void router.navigate({
-      to: '/logout',
-    });
-  };
-
   return (
     <>
       <Meta title="Create Organization" />
       <DottedBackground>
         <div className="flex h-full grow items-center">
-          <Button variant="outline" onClick={handleOnClick} className="absolute right-6 top-6">
+          <Button
+            variant="outline"
+            onClick={() =>
+              void router.navigate({
+                to: '/logout',
+              })
+            }
+            className="absolute right-6 top-6"
+          >
             <LogOutIcon className="mr-2 size-4" /> Sign out
           </Button>
           <Link to="/" className="absolute left-6 top-6">
@@ -33,3 +50,150 @@ export function NewOrgPage(): ReactElement {
     </>
   );
 }
+
+export const CreateOrganizationMutation = graphql(`
+  mutation CreateOrganizationMutation($input: CreateOrganizationInput!) {
+    createOrganization(input: $input) {
+      ok {
+        createdOrganizationPayload {
+          selector {
+            organization
+          }
+          organization {
+            cleanId
+            id
+            cleanId
+          }
+        }
+      }
+      error {
+        inputErrors {
+          name
+        }
+      }
+    }
+  }
+`);
+
+const formSchema = z.object({
+  name: z
+    .string({
+      required_error: 'Organization name is required',
+    })
+    .min(2, {
+      message: 'Name must be at least 2 characters long',
+    })
+    .max(50, {
+      message: 'Name must be at most 50 characters long',
+    })
+    .regex(
+      /^([a-z]|[0-9]|\s|\.|,|_|-|\/|&)+$/i,
+      'Name restricted to alphanumerical characters, spaces and . , _ - / &',
+    ),
+});
+
+export const CreateOrganizationForm = (): JSX.Element => {
+  const [mutation, mutate] = useMutation(CreateOrganizationMutation);
+  const { toast } = useToast();
+  const router = useRouter();
+  const form = useForm<z.infer<typeof formSchema>>({
+    mode: 'onChange',
+    resolver: zodResolver(formSchema),
+    defaultValues: {
+      name: '',
+    },
+    disabled: mutation.fetching,
+  });
+
+  async function onSubmit(values: z.infer<typeof formSchema>) {
+    const mutation = await mutate({
+      input: {
+        name: values.name,
+      },
+    });
+
+    if (mutation.data?.createOrganization.ok) {
+      toast({
+        title: 'Organization created',
+        description: `You are now an admin of "${values.name}" organization.`,
+      });
+      void router.navigate({
+        to: '/$organizationId',
+        params: {
+          organizationId:
+            mutation.data.createOrganization.ok.createdOrganizationPayload.organization.cleanId,
+        },
+      });
+    } else if (mutation.data?.createOrganization.error?.inputErrors?.name) {
+      form.setError('name', {
+        type: 'manual',
+        message: mutation.data.createOrganization.error.inputErrors.name,
+      });
+    } else if (mutation.error) {
+      toast({
+        title: 'Failed to create organization',
+        description: mutation.error.message,
+      });
+    }
+  }
+  return <CreateOrganizationFormContent form={form} onSubmit={onSubmit} />;
+};
+
+type CreateOrganizationFormContentProps = {
+  form: UseFormReturn<z.infer<typeof formSchema>>;
+  onSubmit: (values: z.infer<typeof formSchema>) => void | Promise<void>;
+};
+
+export const CreateOrganizationFormContent = ({
+  form,
+  onSubmit,
+}: CreateOrganizationFormContentProps): JSX.Element => {
+  return (
+    <div className="container w-4/5 max-w-[520px] md:w-3/5">
+      <Form {...form}>
+        <form onSubmit={form.handleSubmit(onSubmit)} className="bg-black">
+          <Card className="w-full">
+            <CardHeader>
+              <CardTitle>Create an organization</CardTitle>
+              <CardDescription>
+                An organization is built on top of <b>Projects</b>. You will become an <b>admin</b>{' '}
+                and don't worry, you can add members later.
+              </CardDescription>
+            </CardHeader>
+            <CardContent>
+              <FormField
+                control={form.control}
+                name="name"
+                render={({ field }) => (
+                  <FormItem>
+                    <FormControl>
+                      <Input placeholder="Organization name" {...field} />
+                    </FormControl>
+                    <FormMessage />
+                  </FormItem>
+                )}
+              />
+            </CardContent>
+            <CardFooter>
+              <Button
+                type="submit"
+                className="w-full"
+                variant="default"
+                disabled={!form.formState.isValid}
+              >
+                {form.formState.isSubmitting ? (
+                  <>
+                    <Spinner className="size-6 text-black" />
+                    <span className="ml-4">Creating...</span>
+                  </>
+                ) : (
+                  'Create Organization'
+                )}
+              </Button>
+            </CardFooter>
+          </Card>
+        </form>
+      </Form>
+    </div>
+  );
+};
